fail if IDS_TOKEN not set

+++ scripts/deploy_cryptomining-tool.sh
@@ -6,8 +6,13 @@ DEVOPS_CONFIG=${DEVOPS_CONFIG:-devops-config}
 VALUES=${DEVOPS_CONFIG}/environments/${ENVIRONMENT}/cryptomining_values.yaml
 COMPONENT_NAME=cryptomining-detector
 
-kubectl -n${CHART_NAMESPACE} delete secret cryptomining-secret
-kubectl -n${CHART_NAMESPACE} create secret generic cryptomining-secret --from-literal=IDS_TOKEN=${IDS_TOKEN}
+if [ -f ${IDS_TOKEN} ]; then
+  kubectl -n${CHART_NAMESPACE} delete secret cryptomining-secret
+  kubectl -n${CHART_NAMESPACE} create secret generic cryptomining-secret --from-literal=IDS_TOKEN=${IDS_TOKEN}
+else
+  echo "IDS_TOKEN is not set"
+  exit 1
+fi
 
 # install yq 
 YQ2_VERSION=2.4.1
